autonomous_trading_system/requirements.txt
# Data Acquisition and Processing
alpaca-trade-api>=2.3.0
polygon-api-client>=1.6.0
pandas>=1.5.3
numpy>=1.24.2
pytz>=2023.3
requests>=2.28.2
websocket-client>=1.5.1
aiohttp>=3.8.4
asyncio>=3.4.3

# Database and Storage
psycopg2-binary>=2.9.5
sqlalchemy>=2.0.7
redis>=4.5.4
pymongo>=4.3.3
timescaledb-python>=0.3.0

# Feature Engineering
ta-lib>=0.4.25
scikit-learn>=1.2.2
statsmodels>=0.13.5
scipy>=1.10.1
featuretools>=1.26.0
tsfresh>=0.20.1

# NLP and Sentiment Analysis
transformers>=4.28.0
torch>=2.0.0
nltk>=3.8.1
spacy>=3.5.2
finbert>=0.2.1

# Machine Learning and Model Training
tensorflow>=2.12.0
keras>=2.12.0
torch>=2.0.0
xgboost>=1.7.5
lightgbm>=3.3.5
catboost>=1.1.1
optuna>=3.1.1
shap>=0.41.0
mlflow>=2.2.2
ray[tune]>=2.3.1

# Trading Strategy
ccxt>=3.0.0
backtrader>=1.9.76.123
pyfolio>=0.9.2
empyrical>=0.5.5
risk-kit>=0.3.0
qgrid>=1.3.1
ipywidgets>=8.0.6

# Monitoring and Alerting
prometheus-client>=0.16.0
grafana-api>=1.0.3
slackclient>=2.9.4
slack-sdk>=3.21.3
pydantic>=1.10.7
fastapi>=0.95.0
uvicorn>=0.21.1

# Utilities
python-dotenv>=1.0.0
click>=8.1.3
tqdm>=4.65.0
loguru>=0.6.0
joblib>=1.2.0
dask>=2023.3.2
distributed>=2023.3.2
boto3>=1.26.115
docker>=6.0.1
kubernetes>=26.1.0

# Development and Testing
pytest>=7.3.1
pytest-cov>=4.1.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.2.0
pre-commit>=3.2.2